CH: Refactoring banner functions
